+++ Source/BugsnagConfiguration.h
@@ -74,10 +74,6 @@ typedef NSDictionary *_Nullable (^BugsnagBeforeNotifyHook)(
  */
 @property(readwrite, retain, nullable) NSString *apiKey;
 /**
- *  The URL used to notify Bugsnag
- */
-@property(readwrite, retain, nullable) NSURL *notifyURL;
-/**
  *  The release stage of the application, such as production, development, beta
  *  et cetera
  */
@@ -142,15 +138,45 @@ BugsnagBreadcrumbs *breadcrumbs;
 @property BOOL autoNotify;
 
 /**
- * Determines whether app sessions should be tracked automatically. By default this value is false.
+ * Determines whether app sessions should be tracked automatically. By default this value is true.
  */
 @property BOOL shouldAutoCaptureSessions;
 
 /**
- * Set the endpoint to which tracked sessions reports are sent. This defaults to https://sessions.bugsnag.com,
- * but should be overridden if you are using Bugsnag On-premise, to point to your own Bugsnag endpoint.
+ * Retrieves the endpoint used to notify Bugsnag of errors
+ *
+ * NOTE: If you want to set this value, you should do so via setEndpointsForNotify:sessions: instead.
+ *
+ * @see setEndpointsForNotify:sessions:
  */
-@property(readwrite, retain, nullable) NSURL *sessionURL;
+@property(readonly, retain, nullable) NSURL *notifyURL;
+
+/**
+ * Retrieves the endpoint used to send tracked sessions to Bugsnag
+ *
+ * NOTE: If you want to set this value, you should do so via setEndpointsForNotify:sessions: instead.
+ *
+ * @see setEndpointsForNotify:sessions:
+ */
+@property(readonly, retain, nullable) NSURL *sessionURL;
+
+/**
+ * Set the endpoints to send data to. By default we'll send error reports to
+ * https://notify.bugsnag.com, and sessions to https://sessions.bugsnag.com, but you can
+ * override this if you are using Bugsnag Enterprise to point to your own Bugsnag endpoint.
+ *
+ * Please note that it is recommended that you set both endpoints. If the notify endpoint is
+ * missing, an assertion will be thrown. If the session endpoint is missing, a warning will be
+ * logged and sessions will not be sent automatically.
+ *
+ * @param notify the notify endpoint
+ * @param sessions the sessions endpoint
+ *
+ * @throws an assertion if the notify endpoint is not a valid URL
+ */
+
+- (void)setEndpointsForNotify:(NSString *_Nonnull)notify
+                     sessions:(NSString *_Nonnull)sessions NS_SWIFT_NAME(setEndpoints(notify:sessions:));
 
 /**
  *  Set user metadata