+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/scenarios/DisabledSessionTrackingScenario.h
@@ -1,0 +1,11 @@
+//
+// Created by Jamie Lynch on 07/06/2018.
+// Copyright (c) 2018 Bugsnag. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+#import "Scenario.h"
+
+
+@interface DisabledSessionTrackingScenario : Scenario
+@end