+++ CHANGELOG.md
@@ -10,6 +10,10 @@ If you use Bugsnag On-Premise, it is now also recommended that you set your noti
 
 * Enable automatic session tracking by default [#286](https://github.com/bugsnag/bugsnag-cocoa/pull/286)
 
+### Bug Fixes
+
+* Handle potential nil content value in RegisterErrorData class [#289](https://github.com/bugsnag/bugsnag-cocoa/pull/289)
+
 ## 5.15.6 (30 May 2018)
 
 ### Bug Fixes