+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Changelog
 
 ## 5.16.0 (TBD)
 
-**IMPORTANT NOTE**: this release alters the behaviour of the notifier to track sessions automatically.
+This release alters the behaviour of the notifier to track sessions automatically.
 A session will be automatically captured on each app launch and sent to [https://sessions.bugsnag.com](https://sessions.bugsnag.com).
 
 If you use Bugsnag On-Premise, it is now also recommended that you set your notify and session endpoints via `config.setEndpoints(notify:sessions:)`. The previous properties used to configure this, `config.notifyURL` and `config.sessionURL`, are now `readonly` and therefore no longer assignable.